+++ CHANGELOG.md
@@ -21,6 +21,7 @@ or any other code that relies on ``Collection.update`` or ``Collection.update_ba
 Using MariaDB might solve this, but it's unclear how the ``JSONField`` performs on MariaDB, and it's not tested against.
 Highly recommended to use Postgres when knowing the ``Document.id`` is important directly after bulk creation.
 
+* Minimal version for Celery is now 5.x.
 * ``global_pipeline_app_label`` and ``global_pipeline_models`` configurations have been renamed
   to ``global_datatypes_app_label`` and ``global_datatype_models``.
 * The ``extractor``, ``depends_on``, ``to_property`` and ``apply_to_resource`` configurations are now